LayoutTests/http/wpt/web-animations/interfaces/AnimationEffectTiming/iterationStart-expected.txt


PASS Test default value 
FAIL Test that changing the iterationStart affects computed timing when backwards-filling assert_approx_equals: expected a number but got a "object"
FAIL Test that changing the iterationStart affects computed timing during the active phase assert_approx_equals: expected a number but got a "object"
FAIL Test that changing the iterationStart affects computed timing when forwards-filling assert_approx_equals: expected a number but got a "object"
PASS Using invalid values should throw TypeError 
